% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addColumn}
\alias{addColumn}
\title{addColumn method for profileMatrix class}
\usage{
addColumn(object, name, col)
}
\arguments{
\item{object}{profileMatrix object}

\item{name}{Character with the name of the column to be add}

\item{col}{Vector of the same lengh as \code{regions(object)}}
}
\value{
Returns a profileMatrix object with the new column added to \code{regions(object)}
}
\description{
This method helps to add a new column to the profile matrix object, is works similarly than using the $ operator over \code{regions(object)}
}
\seealso{
\code{\link{profileMatrix-class}}
}

