% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{bandwidth}
\alias{bandwidth}
\alias{bandwidth,segvis_block-method}
\alias{bandwidth<-}
\alias{bandwidth<-,segvis_block,numeric-method}
\title{bandwidth methods for segvis_block class}
\usage{
bandwidth(object)

bandwidth(object) <- value

\S4method{bandwidth}{segvis_block}(object)

\S4method{bandwidth}{segvis_block,numeric}(object) <- value
}
\arguments{
\item{object}{segvis_block object}

\item{value}{Numeric value}
}
\value{
The smoothing bandwidth of the segvis_block object

A segvis_block object with the bandwidth replcated
}
\description{
bandwidth method returns the smoothing bandwidth of a segvis_block object

bandwidth<- assigns a new smoothing bandwidth to a segvis_block object
}
\examples{
\dontrun{

bandwidth(segvis_block)
bandwidth(segvis_block) <- 101

}
}
\seealso{
\code{\link{segvis_block-class}}

\code{\link{segvis_block-class}}
}

