% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{bandwidth}
\alias{bandwidth}
\alias{bandwidth<-}
\title{bandwidth method for profileMatrix class}
\usage{
bandwidth(object)

bandwidth(object) <- value
}
\arguments{
\item{object}{profileMatrix class}

\item{value}{Numeric value with the bandwidth used to smooth the profile in profileMatrix object}
}
\value{
In case of the get method, the value in the bandwidth slot of the object

In case of the set method, it returns the profileMatrix object with the bandwidth value replaced by value
}
\description{
Get and Set methods for the bandwidth parameter of the profileMatrix class
}
\seealso{
\code{\link{profileMatrix-class}}
}

