% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_buildSegvis.R
\name{buildSegvis}
\alias{buildSegvis}
\title{Builds a segvis object}
\usage{
buildSegvis(name, file, maxBandwidth, chr, fragLen = 0, isPET = FALSE)
}
\arguments{
\item{name}{Name of the segvis object}

\item{file}{Character string with the file name that contain the reads}

\item{maxBandwidth}{Numeric value of the maximum possible value to smooth the profile curve}

\item{chr}{Character vector containing the chromosomes which Segvis is going to process. If the user uses 'human' or 'mouse', then Segvis is going to build the chromosome vector}

\item{fragLen}{Numeric value of the fragment length used to extend the reads}

\item{isPET}{logical, Indicates is the reads come from a PET experiments or a SET experiment}
}
\value{
A segvis object
}
\description{
Constructor for segvis class
}
\examples{
\dontrun{
rn = "sites"
f = "somefile.bam"
maxBw = 1;fl = 200
buildSegvis(regionName = rn,file =f,maxBandwidth = maxBw, fragLen = fl)
}
}

