% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{chr}
\alias{chr}
\alias{chr<-}
\alias{docType}
\alias{methods}
\title{chr methods}
\usage{
chr(object)

chr(object) <- value

\S4method{chr}{segvis}(object)

\S4method{chr}{segvis,character}(object) <- value
}
\arguments{
\item{object}{A \code{segvis} object}

\item{value}{A character vector with the chromosomes to use in the analysis}
}
\value{
A character vector

A segvis object
}
\description{
chr returns a character vector naming the chromosomes used in the analysis

chr<- assigns a logical flag to the segvis object
}
\examples{
\dontrun{

chr(segvis)
chr(segvis) <- c("chr1","chr2",...,"chrX","chrY") ## this is not correct actually

}
}
\seealso{
\code{\link{segvis-class}}, \code{\link{buildSegvis}}
}

