% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{countReads}
\alias{countReads}
\title{countReads method for profile class}
\usage{
countReads(object)
}
\arguments{
\item{object}{profile object}
}
\value{
In case of the get method it returns the constant used to normalize the profile
}
\description{
Counts the number of reads considered  in object
}
\seealso{
\code{\link{profile}}
}

