% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{cover_table}
\alias{cover_table}
\alias{cover_table,segvis_block-method}
\alias{cover_table<-}
\alias{cover_table<-,segvis_block,data.table-method}
\title{cover_table methods for segvis_block class}
\usage{
cover_table(object)

cover_table(object) <- value

\S4method{cover_table}{segvis_block}(object)

\S4method{cover_table}{segvis_block,data.table}(object) <- value
}
\arguments{
\item{object}{segvis_block object}

\item{value}{data.table with the new coverage of the se}
}
\value{
The coverage data.table of a segvis_block object

A segvis_block object with the bandwidth replcated
}
\description{
cover_table method returns coverage table of a segvis_block object

cover_table<- assigns a new coverage table to a segvis_block object
}
\seealso{
\code{\link{segvis_block-class}}

\code{\link{segvis_block-class}}
}

