% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{covers}
\alias{covers}
\alias{covers,SegvizData-method}
\alias{covers}
\title{covers method}
\usage{
covers(object)

\S4method{covers}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{covers} method returns a list of aggregated coverage for
all the aligned reads files considered.
}
\description{
covers returns a list with the \code{coverage} of each file in the
\code{SegvizData} object.
}
\seealso{
\code{\link{SegvizData-class}}
}
