% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extend_to_strand}
\alias{extend_to_strand}
\title{Extend a the reads on a GRanges object to side indicated by it's strand}
\usage{
extend_to_strand(gr, fragLen)
}
\arguments{
\item{gr}{- A GRanges object}

\item{fragLen}{- Numeric value of the size to be of the reads on gr}
}
\value{
A Granges object with extended reads
}
\description{
Extend a the reads on a GRanges object to side indicated by it's strand
}

