% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{file}
\alias{file}
\alias{file,segvis-method}
\alias{file<-}
\alias{file<-,segvis,character-method}
\title{file methods}
\usage{
file(object)

file(object) <- value

\S4method{file}{segvis}(object)

\S4method{file}{segvis,character}(object) <- value
}
\arguments{
\item{object}{A \code{segvis} object}

\item{value}{A character value with the name of bam the file with the reads}
}
\value{
A string

A segvis object
}
\description{
file return the file name used in the segvis object

file<- assigns a string with the file name of the reads to be used
}
\examples{
\dontrun{

file(segvis)
file(segvis) <- "path/to/file/myFile.bam"

}
}
\seealso{
\code{\link{segvis-class}}
}

