% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{file}
\alias{file}
\title{file method for profile class}
\usage{
file(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
character. A character with the name of the file used to create the reads object
}
\description{
Returns the name of the file where the reads are stored
}
\seealso{
\code{\link{profile-class}}
}

