% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{files}
\alias{files}
\alias{files,SegvizData-method}
\alias{files}
\title{files method}
\usage{
files(object)

\S4method{files}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{files} method returns a character vector with the files
used in the \code{SegvizData} object.
}
\description{
files returns a character vectors with the files used in the SegvizData
object.
}
\seealso{
\code{\link{SegvizData-class}}
}
