% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_subset_profile.R
\name{.subset_logical}
\alias{.filter_sb}
\alias{.subset_logical}
\title{Set of functions to subset segvis_block objects}
\usage{
.subset_logical(object, condition_call)

.filter_sb(object, cond)
}
\arguments{
\item{object}{segvis_block object}

\item{condition_call}{name object with the filtering condition}

\item{cond}{Logical Vector with the peaks that satisfy certain condition}
}
\value{
A logical vector with the same length as the number of regions

A segvis_block object with the peaks that such that cond is TRUE
}
\description{
\code{.subset_logical} Function to obtain a logical vector of the elements in regions(object) such that satisfy condition_all

\code{.filter_sb} Filters the segvis_block object using a logical vector as an input, the cond vector can be obtained by using \code{.subset_logical}
}

