% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvisData.R
\docType{methods}
\name{find_summits}
\alias{find_summits}
\alias{find_summits,SegvisData-method}
\alias{find_summits}
\title{find_summits method}
\usage{
find_summits(object, ...)

\S4method{find_summits}{SegvisData}(object, which.file = 1,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{object}{a \code{SegvisData} object.}

\item{...}{Other parameters that may be used by \code{find_summit}.}

\item{which.file}{a numeric value indicating which file is going to be
used to calculate the summits. By default it takes the first file.}

\item{mc.cores}{a numeric value with the number of cores to use,
i.e. at most how many child processes will be run simultaneously.}
}
\value{
The \code{find_summits} method returns a numeric vector with the
genomic coordinate with the summit in the \code{GRanges} according to one
of the files in object.
}
\description{
find_summits returns a numeric vector with the same length of the
\code{GRanges} with the summits according to one of the files.
}
\seealso{
\code{\link{SegvisData-class}}
}
