% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{frag_len}
\alias{frag_len}
\alias{frag_len,SegvizData-method}
\alias{frag_len}
\title{frag_len method}
\usage{
frag_len(object)

\S4method{frag_len}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{frag_len} method returns a numeric vector with the fragment
length used to extend the 5' ends of the the aligned reads. If the reads
were paired ended, then a value of \code{NA} is assigned, but the parameter is
not used.
}
\description{
frag_len returns a numeric vector with the fragment length used to extend
the 5' ends of the aligned reads.
}
\seealso{
\code{\link{SegvizData-class}}
}
