% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{fwd_covers}
\alias{fwd_covers}
\alias{fwd_covers,SegvizData-method}
\title{fwd_covers method}
\usage{
fwd_covers(object)

\S4method{fwd_covers}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{fwd_covers} method returns a list of aggregated coverage
for all the aligned reads files considered, by using the reads in the forward
strand exclusively.
}
\description{
fwd_covers returns a list with the \code{coverage} of each file in the
\code{SegvizData} object, calculated exclusively with the forward strand
reads.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
fwd_covers(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

