% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{getCoverage}
\alias{getCoverage}
\alias{getCoverage,segvis,numeric-method}
\title{getCoverage method for profile class}
\usage{
getCoverage(object, mc)

\S4method{getCoverage}{segvis,numeric}(object, mc = 8)
}
\arguments{
\item{object}{segvis object}

\item{mc}{numeric, the number of cores used with parallel}
}
\value{
segvis object
}
\description{
Calculate  the coverage using the reads matched by the matchReads method. It returns a Rle object for each region.
}
\examples{
\dontrun{

getCoverage(segvis_object,mc=8)

}
}
\seealso{
\code{\link{loadReads}}, \code{\link{matchReads}} and \code{\link{findSummit}}
}

