% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getCoverage}
\alias{getCoverage}
\title{getCoverage method for profile class}
\usage{
getCoverage(object, mc)
}
\arguments{
\item{object}{profile object}

\item{mc}{numeric, the number of cores used with parallel}
}
\value{
profile object
}
\description{
Build the coverage using the reads matched by the matchReads method. It returns a Rle object for each region.
}
\examples{
\dontrun{ getCoverage(profile_object,mc=8)}
}
\seealso{
\code{\link{loadReads}} and \code{\link{matchReads}}
}

