% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{isPET}
\alias{isPET}
\alias{isPET<-}
\title{isPET methods}
\usage{
isPET(object)

isPET(object) <- value

\S4method{isPET}{segvis}(object)

\S4method{isPET}{segvis,logical}(object) <- value
}
\arguments{
\item{object}{A \code{segvis} object}

\item{value}{A logical flag that represents wheter the reads to use are Paired Ended or not}
}
\value{
A logical flag

A segvis object
}
\description{
isPET returns a logical flag wheter the reads to consider are Paired Ended or not

isPET<- assigns a logical flag to the segvis object
}
\examples{
\dontrun{

isPET(segvis)
isPET(segvis) <- TRUE

}
}
\seealso{
\code{\link{segvis-class}}
}

