% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{joinProfiles}
\alias{joinProfiles}
\alias{joinProfiles,segvis,numeric,numeric-method}
\title{joinProfiles method for segvis class}
\usage{
joinProfiles(object, bw, mc)

\S4method{joinProfiles}{segvis,numeric,numeric}(object, bw, mc = 8)
}
\arguments{
\item{object}{segvis object}

\item{bw}{Numeric value with the bandwidth used to smooth the coverage profiles, must be and odd number less or equal than the maxBandwidth}

\item{mc}{Numeric value with the number of cores used with parallel}
}
\value{
A data.table object
}
\description{
Joins all the coverages into a data.table with columns chr|match|coord|tagCounts
}
\examples{
\dontrun{

joinProfiles(segvis_object,bw = 151,mc=4)

}
}
\seealso{
\code{\link{segvis_block-class}},\code{\link{segvis-class}}
}

