% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{match-class}
\alias{match-class}
\title{match class}
\description{
Contains the match of a set a reads of a ChIP - Seq experiment given a set
of regions in the genome. We are obtaining a match for each strand with the
match_reads.cpp function
}
\section{Slots}{

\describe{
\item{\code{match1}}{- List of positions of the reads in list that match with each region and have + strand}

\item{\code{match2}}{- List of positions of the reads in list that match with each region and have - strand}
}}
\seealso{
\code{\link{match1}}, \code{\link{match2}} and matchReads
}

