% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{match1}
\alias{match}
\alias{match1}
\alias{match2}
\title{match1}
\usage{
match1(object, ...)

match2(object, ...)
}
\arguments{
\item{match}{object}

\item{match}{object}
}
\value{
list The match of the reads with "+" strand

list The match of the reads with "-" strand
}
\description{
match1

match2
}

