% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{match2}
\alias{match2}
\title{match2}
\usage{
match2(object)
}
\arguments{
\item{match}{object}
}
\value{
list The match of the reads with "-" strand
}
\description{
Get method for match class with "-" strand
}
\details{
This methods returns a list with the indexes in the "-" strand GRangesList
}
\seealso{
\code{\link{reads2}} and \code{\link{match1}}
}

