% Generated by roxygen2 (4.0.1): do not edit by hand
\name{match_reads}
\alias{match_reads}
\title{Creates a list with the matched reads for each regions}
\usage{
match_reads(regionStart, regionEnd, readStart, readEnd, strand, fraglen)
}
\arguments{
\item{regionStart}{- start(regions) from GRanges}

\item{regionEnd}{- end(regions) from GRanges}

\item{readStart}{- start(reads) from GRanges}

\item{readEnd}{- end(reads) from GRanges}

\item{strand}{- as.character(strand(reads)) form GRanges}

\item{fraglen}{- Numeric value, fragment length}
}
\value{
position in reads vector
}
\description{
Creates a list with the matched reads for each regions
}

