% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{matchF}
\alias{matchF}
\title{matchF method for profile and match classes}
\usage{
matchF(object)
}
\arguments{
\item{match}{object}
}
\value{
In case of the get method returns the reads of the forward strand
}
\description{
Get method for match class with "+" strand
}
\seealso{
\code{\link{match-class}}
}

