% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{matchR}
\alias{matchR}
\title{matchR method for profile and match clases}
\usage{
matchR(object)
}
\arguments{
\item{match}{object}
}
\value{
list The match of the reads with "-" strand
}
\description{
Get method for match class with "-" strand
}
\seealso{
\code{\link{match-class}}
}

