% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{matchReads}
\alias{matchReads}
\alias{matchReads,segvis,numeric-method}
\title{matchReads methods for segvis class}
\usage{
matchReads(object, mc)

\S4method{matchReads}{segvis,numeric}(object, mc = 8)
}
\arguments{
\item{object}{segvis object}

\item{mc}{numeric, the number of cores used with parallel}
}
\value{
segvis object
}
\description{
Match the reads to the extended regions stored in the object slots.
}
\examples{
\dontrun{

matchReads(segvis_object,mc=8)

}
}
\seealso{
\code{\link{readsF}}, \code{\link{readsR}} and \code{\link{reads-class}}
}

