% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{maxBandwidth}
\alias{maxBandwidth}
\alias{maxBandwidth,segvis-method}
\alias{maxBandwidth<-}
\alias{maxBandwidth<-,segvis,numeric-method}
\title{maxBandwidth methods}
\usage{
maxBandwidth(object)

maxBandwidth(object) <- value

\S4method{maxBandwidth}{segvis}(object)

\S4method{maxBandwidth}{segvis,numeric}(object) <- value
}
\arguments{
\item{object}{A \code{segvis} object}

\item{value}{A numeric value representing the maximum bandwidth, it need to be an odd number}
}
\value{
A integer value

A segvis object
}
\description{
maxBandwidth returns a numeric value representing the possible bandwidths upper bound to smooth the coverage profiles.

maxBandwidth<- assigns an integer value the bandwidth upper bound to use when smoothing the profiles in segvis_block object
}
\examples{
\dontrun{

maxBandwidth(segvis)
maxBandwidth(segvis) <- 201

}
}
\seealso{
\code{\link{segvis-class}}
}

