% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{meanProfile}
\alias{meanProfile}
\title{meanProfile method for profileMatrix class}
\usage{
meanProfile(object, ...)
}
\arguments{
\item{object}{profileMatrix object}

\item{trim}{Numeric. By default is set to zero. If trim =0, it calculated the row-wise mean, if trim >= 0.5 then is calculates the row-wise median, otherwise it calculate a row-wise trimmed mean.}
}
\value{
Numeric vector
}
\description{
This method calculates the "average" profile of the matrix respect to their rows. As in the mean function it is possible to set a trim parameter
}
\seealso{
\code{\link{profileMatrix-class}}
}

