% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{mergeList}
\alias{mergeList}
\title{mergeList method for profileMatrixList class}
\usage{
mergeList(object)
}
\arguments{
\item{object}{profileMatrixList with the profile matrices to be merged}
}
\value{
A profileMatrix object with the regions and matrices merged
}
\description{
This methods returns a profileMatrix class object which contains the regions and coverage curves of the list merged
}
\details{
In this case, we are going to merge by consider the priority implied by the order in list. Which means if there is an overlap between a region of object[[1]] with a region of object[[2]], the algorithm is going to choose the region of object[[1]] and so on.
}
\seealso{
\code{\link{profileMatrixList-class}}
}

