% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-reads.R
\docType{methods}
\name{readsF}
\alias{readsF}
\alias{readsF<-}
\alias{readsR}
\alias{readsR<-}
\title{Get and set method of reads objects}
\usage{
readsF(object)

readsF(object) <- value

readsR(object)

readsR(object) <- value

\S4method{readsF}{reads}(object)

\S4method{readsR}{reads}(object)

\S4method{readsF}{reads,list}(object) <- value

\S4method{readsR}{reads,list}(object) <- value
}
\arguments{
\item{object}{Either a \code{segvis} o \code{reads} object}

\item{value}{A data.table converted from a \code{GRanges} object}
}
\description{
\code{readsF} operates on the forward reads, while \code{readsR} operates on the backward reads
}

