% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{name}
\alias{file}
\alias{file<-}
\alias{fragLen}
\alias{fragLen<-}
\alias{matchF}
\alias{matchR}
\alias{maxBandwidth}
\alias{maxBandwidth<-}
\alias{name}
\alias{name<-}
\alias{profileCurve}
\alias{readsF}
\alias{readsR}
\alias{regions}
\alias{regions<-}
\title{Get and set methods for segvis class}
\usage{
name(object)

name(object) <- value

\S4method{name}{segvis}(object)

\S4method{regions}{segvis}(object)

\S4method{file}{segvis}(object)

\S4method{maxBandwidth}{segvis}(object)

\S4method{fragLen}{segvis}(object)

\S4method{readsF}{segvis}(object)

\S4method{readsR}{segvis}(object)

\S4method{matchF}{segvis}(object)

\S4method{matchR}{segvis}(object)

\S4method{profileCurve}{segvis}(object)

\S4method{name}{segvis,character}(object) <- value

\S4method{regions}{segvis,GRanges}(object) <- value

\S4method{file}{segvis,character}(object) <- value

\S4method{maxBandwidth}{segvis,numeric}(object) <- value

\S4method{fragLen}{segvis,numeric}(object) <- value
}
\arguments{
\item{object}{An object of class segvis}

\item{value}{The class of this parameter depends on the function, for \code{file<-} and \code{name<-} is a string; for \code{fragLen<-} and  \code{maxBandwidth} is a numeric value and for \code{regions<-} is a GRanges object}
}
\value{
Returns the value with the same name as the function

Returns a segvis object with the value indicated by the functions being updated with value
}
\description{
This methods determine how to get and set attributes dor the segvis class. The structure is similar to the one use by bioconductor objects
}
\seealso{
\code{\link{segvis-class}}
}

