% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{addColumn}
\alias{addColumn}
\title{addColumn method for segvis_block class}
\usage{
addColumn(object, name, col)
}
\arguments{
\item{object}{segvis_block object}

\item{name}{Character with the name of the column to be add}

\item{col}{Vector of the same lengh as \code{regions(object)}}
}
\value{
Returns a segvis_block object with the new column added to \code{regions(object)}
}
\description{
This method helps to add a new column to the profile matrix object, is works similarly than using the $ operator over \code{regions(object)}
}
\seealso{
\code{\link{segvis_block-class}}
}

