% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{cover_table}
\alias{bandwidth}
\alias{bandwidth<-}
\alias{cover_table}
\alias{cover_table<-}
\alias{name}
\alias{regions}
\title{cover_table method for segvis_block class}
\usage{
cover_table(object)

cover_table(object) <- value

bandwidth(object)

bandwidth(object) <- value

\S4method{name}{segvis_block}(object)

\S4method{regions}{segvis_block}(object)

\S4method{cover_table}{segvis_block}(object)

\S4method{bandwidth}{segvis_block}(object)

\S4method{name}{segvis_block,character}(object) <- value

\S4method{regions}{segvis_block,GRanges}(object) <- value

\S4method{cover_table}{segvis_block,data.table}(object) <- value

\S4method{bandwidth}{segvis_block,numeric}(object) <- value
}
\arguments{
\item{object}{An object of class segvis_block}

\item{value}{The class of this parameter depends on the function, \code{name<-} is a string; for \code{bandwidth<-} and  \code{normConst} is a numeric value; for \code{regions<-} is a GRanges object and for \code{cover_table<-} is a data.table}
}
\value{
Returns the value with the same name as the function

Returns a segvis object with the value indicated by the functions being updated with value
}
\description{
This methods determine how to get and set attributes for the segvis_block class. The structure is similar to the one use by bioconductor objects
}
\seealso{
\code{\link{segvis-class}}
}

