% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{normalize}
\alias{normalize}
\title{normalize method for segvis_block class}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{segvis_block object}

\item{value}{A positive number or a missing value. If is missing is going to normalize as if they were 1 million reads in the experiment}

\item{base}{A positive integer considering the base used to normalize the tagCounts}
}
\value{
A segvis_block object where the cover_table contains normalized tagCounts
}
\description{
This method normalizes the segvis_block tagCounts by multiplying it times value.
It normalizes the tagCounts by multiplying it times base / normConst
}
\seealso{
\code{\link{segvis_block-class}}
}

