% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{subset}
\alias{subset}
\title{subset method for segvis_block class}
\usage{
subset(object, condition)

\S4method{subset}{segvis_block}(object, condition)
}
\arguments{
\item{object}{segvis_block object}

\item{condition}{This is an expression considering the characteristics taht the subset need to satisfy}
}
\value{
Returns a segvis_block object with the same parameters as object except regions and profileMat which are filtered to satisfy the conditions on condition.
}
\description{
This method works similarly to the subset of IRanges, GenomicRanges, GenomicAlignments, etc. Althought it doesn't consider the select parameter.
}
\seealso{
\code{\link{segvis_block-class}}
}

