% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block_list.R
\docType{methods}
\name{plot_data}
\alias{plot_data}
\title{plot_profiles method for segvis_block_list class}
\usage{
plot_data(object, condition, coord, FUN, ..., mc)
}
\arguments{
\item{object}{segvis_block_list object}

\item{condition}{The condition used to filter the profileMatrix objects, if there is no condition then it doesn't filter the objects}

\item{coord}{Numeric vector representing the coordinates used in the x-axis of the plot, by default considers the natural index \code{1:ncol(segvis_block)}}

\item{FUN}{Function used to summarize the profiles}

\item{...}{Additional arguments of \code{FUN}}
}
\value{
A data.table containing the data ready to plot by ggplot
}
\description{
This method returns the ggplot object used later to print the plot
}
\seealso{
\code{\link{segvis_block_list-class}}
}

