% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R, R/methods-segvis_block.R
\docType{methods}
\name{name}
\alias{name}
\alias{name,segvis-method}
\alias{name,segvis_block-method}
\alias{name<-}
\alias{name<-,segvis,character-method}
\alias{name<-,segvis_block,character-method}
\title{name methods}
\usage{
name(object)

name(object) <- value

\S4method{name}{segvis}(object)

\S4method{name}{segvis,character}(object) <- value

\S4method{name}{segvis_block}(object)

\S4method{name}{segvis_block,character}(object) <- value
}
\arguments{
\item{object}{A \code{segvis} object}

\item{value}{A character with the name of the object}
}
\value{
A string with the name of the object

A segvis object
}
\description{
name returns a string with the name of the segvis object

name<- assigns a new a name to the segvis object
}
\examples{
\dontrun{

name(segvis)
name(segvis) <- "my_regions"

}
}
\seealso{
\code{\link{segvis-class}}
}

