% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{name}
\alias{name}
\alias{name<-}
\title{name get and set methods for profile and profileMatrix}
\usage{
name(object)

name(object) <- value
}
\arguments{
\item{object}{Either a profile or profileMatrix object}

\item{value}{character}
}
\value{
Returns the name of the object

Returns and profile (or profileMatrix) object with the name parameter replaced by value
}
\description{
Generic method for both profile and profileMatrix classes that returns the name of the object
}
\seealso{
\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

