% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,segvis_block,ANY-method}
\alias{normalize,segvis_block,numeric-method}
\title{normalize method for segvis_block class}
\usage{
normalize(object, base)

\S4method{normalize}{segvis_block,numeric}(object, base)

\S4method{normalize}{segvis_block,ANY}(object)
}
\arguments{
\item{object}{segvis_block object}

\item{base}{Numeric value, indicating to which scale the profiles are gonna be scaled}
}
\description{
This method normalizes the segvis_block tagCounts by multiplying it times value. It normalizes the tagCounts by multiplying it times base / normConst
}
\examples{
\dontrun{

normalize(segvis_block)
normalize(segvis_block,base = 1e9) ## rpkm

}
}
\seealso{
\code{\link{segvis_block-class}}
}

