% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{normalize.matrix}
\alias{normalize.matrix}
\title{normalize.matrix method for profileMatrix class}
\usage{
normalize.matrix(object, ...)
}
\arguments{
\item{object}{profileMatrix object}

\item{value}{A positive number or a missing value. If is missing is going to normalize as if they were 1 million reads in the experiment}
}
\value{
A profileMatrix object where the profileMat slot has been normalized
}
\description{
This method normalizes the profile matrix by multiplying it times value
}
\seealso{
\code{\link{profileMatrix-class}}
}

