% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{normalize}
\alias{normalize}
\title{normalize method for segvis_block class}
\usage{
normalize(object, ...)
}
\arguments{
\item{object}{segvis_block object}

\item{...}{Remaining \code{normalize} parameters. \code{value} and \code{base}}
}
\description{
This method normalizes the segvis_block tagCounts by multiplying it times value. It normalizes the tagCounts by multiplying it times base / normConst
}
\examples{
\dontrun{

normalize(segvis_block)
normalize(segvis_block,base = 1e9) ## rpkm

}
}
\seealso{
\code{\link{segvis_block-class}}
}

