% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{normConst}
\alias{normConst}
\alias{normConst,segvis_block-method}
\alias{normConst<-}
\alias{normConst<-,segvis_block,numeric-method}
\title{normConst methods for segvis_block class}
\usage{
normConst(object)

normConst(object) <- value

\S4method{normConst}{segvis_block}(object)

\S4method{normConst}{segvis_block,numeric}(object) <- value
}
\arguments{
\item{object}{segvis_block object}

\item{value}{Numeric value}
}
\value{
The normalizing constant of the segvis_block object

A segvis_block object with the normConst replcated
}
\description{
normConst method returns the normalizing constant of a segvis_block object

normConst<- assigns a new normalizing constant to a segvis_block object
}
\examples{
\dontrun{

normConst(segvis_block)
normConst(segvis_block) <- 100
normConst(segvis_block) <- countReads(segvis)

}
}
\seealso{
\code{\link{segvis_block-class}}

\code{\link{segvis_block-class}}
}

