% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{normConst}
\alias{normConst}
\alias{normConst<-,ANY-methods}
\title{normConst method for segvis_block class}
\usage{
normConst(object)

normConst(object) <- value

\S4method{normConst}{segvis_block}(object)

\S4method{normConst}{segvis_block,numeric}(object) <- value
}
\arguments{
\item{object}{segvis_block object}

\item{value}{Numeric normalizing constant value}

\item{object}{segvis_block object}
}
\value{
A numeric value

A segvis_block object
}
\description{
This method returns the normalizing constant of a \code{segvis_block} object

This method assigns a new normalizing constant to a \code{segvis_block} object
}
\examples{
\dontrun{

normConst(segvis_block)
}
\dontrun{

normConst(segvis_block) <- countReads(segvis)
normConst(segvis_block) <- 1e3

}
}
\seealso{
\code{\link{segvis_block-class}}

\code{\link{segvis_block-class}}
}

