% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{overlap_matrix}
\alias{overlap_matrix}
\alias{overlap_matrix,SegvizData,character-method}
\title{overlap_matrix method}
\usage{
overlap_matrix(object, bedfiles, ...)

\S4method{overlap_matrix}{SegvizData,character}(object, bedfiles,
  colnames = basename(bedfiles))
}
\arguments{
\item{object}{a \code{SegvizData} object.}

\item{bedfiles}{a character vector with the files in bed fortmat.}

\item{...}{Any other additional parameters that \code{overlap_matrix} may need.}

\item{colnames}{a character vector with the names of the bed files. The
default value for this is \code{basename(bedfiles)}.}
}
\value{
The \code{overlap_matrix} method returns an overlap matrix where each
entry indicates if the row-region overlaps any of the intervals in the
bedfiles.
}
\description{
overlap_matrix returns a matrix of zeros and ones indicating if the regions
in the \code{GRanges} object overlap a region in the respective bedfile
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
bedfiles = list.files(system.file("extdata","example",package = "Segvis"),
  pattern = "narrow",full.names = TRUE)[-1]
  overlap_matrix(segviz,bedfiles,c("dnase1","dnase2"))

}
\seealso{
\code{\link{SegvizData-class}}
}

