% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plot.profiles}
\alias{plot.profiles}
\title{plot.profile method for profileMatrixList class}
\usage{
plot.profiles(object, ...)
}
\arguments{
\item{object}{profileMatrixList object}

\item{condition}{The condition used to filter the profileMatrix objects, if there is no condition then it doesn't filter the objects}

\item{coord}{Numeric vector representing the coordinates used in the x-axis of the plot, by default considers the natural index \code{1:ncol(profileMatrix)}}

\item{trim}{Numeric. By default is set to zero. If trim =0, it calculated the row-wise mean, if trim >= 0.5 then is calculates the row-wise median, otherwise it calculate a row-wise trimmed mean.}
}
\value{
The ggplot object used to print the plot
}
\description{
This method returns the ggplot object used later to print the plot
}
\seealso{
\code{\link{profileMatrixList-class}}
}

