% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_segvis_block_list.R
\name{plot_data}
\alias{plot_data}
\title{Generates the data to be plotted}
\usage{
plot_data(object, condition, coord = NULL, mc, FUN, ...)
}
\arguments{
\item{object}{segvis_block_list object}

\item{condition}{This is an expression used to filter the \code{segvis_block} objects, if there is no condition, then it doesn't filter the objects}

\item{coord}{Numeric vector representing the coordinates in the a-axis of the plot, by default uses the natural index \code{1:ncol(segvis_block)}}

\item{mc}{Numeber of cores used by parallel}

\item{FUN}{function used to summarize the profiles}

\item{...}{Additional arguments of \code{FUN}}
}
\value{
A data.table object
}
\description{
This function generates a data.table with the profiles that would be in the profile plot
}
\examples{
\dontrun{

 plot_data(segvis_block_list,mc = 8, mean, trim = .05)

 ## if the segvis_block_list have a column called row
 plot_data(segvis_block_list,condition = row == 1,mc = 8,FUN = function(x)x)

}
}
\seealso{
\code{\link{segvis_block_list-class}},\code{\link{plot_profiles}}
}

