% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap,SegvizData-method}
\title{plot_heatmap method}
\usage{
plot_heatmap(object, ...)

\S4method{plot_heatmap}{SegvizData}(object, which.cluster = 1,
  dist_method = "euclidean", clust_method = "complete",
  nameFiles = basename(files(object)), type = "aggr", base = 1e+06,
  mc.cores = getOption("mc.cores", 2L), ...)
}
\arguments{
\item{object}{a \code{SegvizData} object.}

\item{which.cluster}{an integer indicating which signal is going to be used to
cluster the profiles. By default uses the first file loaded.}

\item{dist_method}{the distance measure to be used. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{clust_method}{the agglomeration method to be used. This should be
one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA),
"mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{nameFiles}{a character vector with the shortened names that are going
to be used in the plot.}

\item{type}{a character value indicating which strand to use. By default
uses the aggregate coverage between both strands.}

\item{base}{a numeric value indicating the number of aligned reads to which
the signal is going to be normalized. The default value is \code{1e6}.
\code{DT_profile} always normalizes the signal.}

\item{mc.cores}{a numeric value indicating the number of multi-cores to be
used.}
}
\value{
The \code{plot_heatmap} plots a heatmap that summarizes the signal
profiles of all files in the \code{SegvizData} object.
}
\description{
plot_heatmap plots a heatmap that summarizes the signal profiles of all files
in the \code{SegvizData} object as a function of the distance to the
center of the regions.
}
\seealso{
\code{\link{SegvizData-class}}
## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
plot_heatmap(segviz)
}

