% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block_list.R
\docType{methods}
\name{plot_profiles}
\alias{plot_profiles}
\alias{plot_profiles,ANY-method}
\alias{plot_profiles-methods}
\title{plot_profiles method for segvis_block_list class}
\usage{
plot_profiles(object, condition, coord, FUN, ..., mc)
}
\arguments{
\item{object}{segvis_block_list object}

\item{condition}{The condition used to filter the profileMatrix objects, if there is no condition then it doesn't filter the objects}

\item{coord}{Numeric vector representing the coordinates used in the x-axis of the plot, by default considers the natural index \code{1:ncol(segvis_block)}}

\item{FUN}{Function used to summarize the profiles}

\item{...}{Additional arguments of \code{FUN}}

\item{mc}{Number of cores used by parallel}
}
\value{
The ggplot object used to print the plot
}
\description{
This method returns the ggplot object used later to print the plot
}
\examples{
\dontrun{

 plot_profiles(segvis_block_list, mean, trim = .05,mc = 8)

 ## if the segvis_block_list have a column called row
 plot_profiles(segvis_block_list, FUN = function(x)x,condition = row == 1,mc = 8)

}
}
\seealso{
\code{\link{segvis_block_list-class}},\code{\link{plot_data}}
}

