% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_segvis_block_list.R
\docType{methods}
\name{plot_profiles}
\alias{plot_profiles}
\title{Plots the coverage profiles for the segvis_block_list}
\usage{
plot_profiles(object, condition, coord = NULL, mc, FUN, ...)
}
\arguments{
\item{object}{segvis_block_list object}

\item{condition}{This is an expression used to filter the \code{segvis_block} objects, if there is no condition, then it doesn't filter the objects}

\item{coord}{Numeric vector representing the coordinates in the a-axis of the plot, by default uses the natural index \code{1:ncol(segvis_block)}}

\item{mc}{Numeber of cores used by parallel}

\item{FUN}{function used to summarize the profiles}

\item{...}{Additional arguments of \code{FUN}}
}
\value{
The ggplot object used to print the plot
}
\description{
This function generates a ggplot object with the desired coverage plots
}
\examples{
\dontrun{

 plot_profiles(segvis_block_list, mean, trim = .05,mc = 8)

 ## if the segvis_block_list have a column called row
 plot_profiles(segvis_block_list, FUN = function(x)x,condition = row == 1,mc = 8)

}
}
\seealso{
\code{\link{segvis_block_list-class}},\code{\link{plot_data}}
}

