% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{profile-class}
\alias{profile-class}
\title{profile class}
\description{
Contains all the information necessary for the calculation of profile curves.
}
\section{Slots}{

\describe{
\item{\code{name}}{- Character with the name of the profiles}

\item{\code{regions}}{- List with the regions for which the coverage is going to be calculated}

\item{\code{files}}{- Character with the files which contains the reads}

\item{\code{fileFormat}}{- Character with the file format used for the reads}

\item{\code{maxBandwidth}}{- Numeric - maximum bandwidth accepted when smoothing profiles. Must be odd}

\item{\code{fragLen}}{- Numeric fragment length to extend the reads}

\item{\code{remChr}}{- Character - Vector with the chromosomes to be ignored}

\item{\code{readsList}}{- List of reads objects}

\item{\code{matchList}}{- List of match objects}

\item{\code{profileCuve}}{- RleList - For each region, there is a Rle object}

\item{\code{.haveRegions}}{- logical - Indicates if the object have the regions loaded}

\item{\code{.haveReads}}{- logical - Indicates if the object have the reads loaded}

\item{\code{.readsMatched}}{- logical - Indicates if the read have been matched to the regions}

\item{\code{.coverageCalculated}}{- logical - Indicates if the coverage has been calculated]}
}}

