% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remChr}
\alias{matchList}
\alias{readsList}
\alias{remChr}
\title{remChr}
\usage{
remChr(object, ...)

readsList(object, ...)

matchList(object, ...)
}
\arguments{
\item{profile}{object}

\item{profile}{object}

\item{profile}{object}
}
\value{
character. A character vector representing the chromosomes to be ignored

list. A list made off reads objects. One list for each replicate. Must coincide with size of bedfiles

list. A list made off match objects. One list for each replicate. Must coincide with size of bedfiles
}
\description{
remChr

readsList

matchList
}

