% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{profileCurve}
\alias{buildProfileMat}
\alias{getCoverage}
\alias{loadReads}
\alias{matchReads}
\alias{profileCurve}
\alias{setFragLen}
\alias{setMaxBandwidth}
\alias{setRemChr}
\title{profileCurve}
\usage{
profileCurve(object, ...)

setMaxBandwidth(object, newMaxBandwidth)

setFragLen(object, newFragLen)

setRemChr(object, newRemChr)

loadReads(object, mc)

matchReads(object, mc)

getCoverage(object, mc)

buildProfileMat(object, bw, mc)
}
\arguments{
\item{profile}{object}

\item{profile}{object}

\item{newMaxBandwidth}{Numeric value, must be odd and greater or equal than one}

\item{profile}{object}

\item{newFragLen}{Numeric value, must be greater of equal to zero}

\item{profile}{object}

\item{newRemChr}{Character vector, with the chromosomes to be removed}

\item{profile}{object}

\item{mc}{numeric, the number of cores used with parallel}

\item{profile}{object}

\item{mc}{numeric, the number of cores used with parallel}

\item{profile}{object}

\item{mc}{numeric, the number of cores used with parallel}

\item{profile}{object}

\item{bw,}{the bandwidth used to smooth the profiles}

\item{mc,}{the number of cores used with parallel}
}
\value{
RleList. A list made of an Rle object for each region

profile object

profile object

profile object

profile object

profile object

profile object

list object
}
\description{
profileCurve

setMaxBandwidth

setFragLen

setRemChr

loadReads

matchReads

getCoverage

buildProfileMat
}

