% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Profile}
\alias{Profile}
\title{Create a profile object}
\usage{
Profile(regionName, files, fileF, maxBw, fl, rem_Chr)
}
\arguments{
\item{regionName}{- Name of the profile object}

\item{files}{- Character vector with files that contain the reads}

\item{fileFormat}{- Character vector with the file format used}

\item{fl}{- Numeric value of the fragment length used to extend the reads}

\item{maxBw}{- Numeric value of the maximum possible value to smooth the profile curve}

\item{rem_Chr}{- Character Vector with the chromosomes to be ignored}
}
\value{
A profile object
}
\description{
Create a profile object
}

