% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{profileCurve}
\alias{profileCurve}
\title{profileCurve methods for profile class}
\usage{
profileCurve(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
RleList. A list made of an Rle object for each region
}
\description{
Returns the coverage for each region considered
}
\seealso{
\code{\link{profile-class}} and \code{\link{getCoverage}}
}

