% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{profileMat}
\alias{profileMat}
\title{profileMat method for profileMatrix class}
\usage{
profileMat(object)
}
\arguments{
\item{object}{profileMatrix object}
}
\value{
matrix object. A matrix with a the profile for each region
}
\description{
Get method of the profile matrix for the class
}
\seealso{
\code{\link{buildProfileMatrix}}, \code{\link{setProfileMat}} and ProfileMatrix
}

