% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{profileMatrix-class}
\alias{profileMatrix-class}
\title{profileMatrix class description}
\description{
Contains a matrix with the individual coverage for each region
}
\section{Slots}{

\describe{
\item{\code{name}}{- Character with the name of the profiles}

\item{\code{regions}}{- GRanges with the regions}

\item{\code{profileMat}}{- Matrix - Actual profile matrix}

\item{\code{bandwidth}}{- Numeric value used to smooth the individual coverages}

\item{\code{normConst}}{- Numeric Normalizing constant}

\item{\code{.isScaled}}{- Logical representing if the profile matrix is scaled}
}}

