% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{profMat}
\alias{bandwidth}
\alias{name}
\alias{normConst}
\alias{profMat}
\alias{profileMatrix}
\alias{regions}
\alias{setBandwidth}
\alias{setName}
\alias{setNormConst}
\alias{setProfMat}
\alias{setRegions}
\title{profMat}
\usage{
profMat(object)

bandwidth(object)

normConst(object)

setProfMat(object, newProfMatrix)

setBandwidth(object, newBandwidth)

setNormConst(object, newNormConst)
}
\arguments{
\item{object.}{profileMatrix}

\item{object.}{profileMatrix}

\item{object.}{profileMatrix}

\item{object.}{profileMatrix}

\item{newProfMat}{matrix object. A matrix with a the profile for each region}

\item{object.}{profileMatrix}

\item{newBandwidth.}{Numeric value with the bandwidth used to smooth the profile in profileMatrix object}

\item{object.}{profileMatrix}

\item{newNormConst}{Numeric value with the constant used to normalize the profile}
}
\value{
matrix object. A matrix with a the profile for each region

numeric. The bandwidth used to smooth the profile in profileMatrix object

numeric. The constant used to normalize the profile

profileMatrix object

profileMatrix object

profileMatrix object
}
\description{
profMat

bandwidth

normConst

setProfMat

setBandwidth

setNormConst
}

