% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ProfileMatrix}
\alias{ProfileMatrix}
\title{Create a profileMatrix object}
\usage{
ProfileMatrix(object, bw, mc)
}
\arguments{
\item{object}{profile object}

\item{bw}{Numeric used to smooth the profile curves}

\item{mc}{Numeric, the number of cores used with parallel}
}
\value{
a profileMatrix object
}
\description{
Constructor for profileMatrix class, it takes both name and region from \code{object}
}
\examples{
\dontrun{ProfileMatrix(profile_object,bw = 151,mc=8)}
}

