% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{profiles}
\alias{profiles}
\title{profiles methods for segvis class}
\usage{
profiles(object)

\S4method{profiles}{segvis}(object)
}
\arguments{
\item{object}{A segvis object}
}
\value{
RleList. A list made of an Rle object for each region
}
\description{
Returns the coverage for each region considered
}
\examples{
\dontrun{

profiles(segvis)

}
}
\seealso{
\code{\link{segvis-class}} and \code{\link{getCoverage}}
}

