% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utilities.R
\name{readBedFile}
\alias{readBedFile}
\title{readBedFile}
\usage{
readBedFile(file)
}
\arguments{
\item{file}{a character with the increased bed file location.}
}
\value{
A \code{GRanges} object.
}
\description{
Reads an increased bed file and adds the output into a \code{GRanges}
object.
}
\references{
\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}
}
