% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{reads-class}
\alias{reads-class}
\title{reads class description}
\description{
Contains the reads obtained in a ChIP - seq experiment separated by strand and then by chromosome.
}
\section{Slots}{

\describe{
\item{\code{readsF}}{- GRangesList of the reads of the ChIP - Seq experiment that have + strand}

\item{\code{readsR}}{- GRangesList of the reads of the ChIP - Seq experiment that have - strand}
}}
\seealso{
\code{\link{readsF}}, \code{\link{readsR}} and \code{\link{loadReads}}
}

