% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reads1}
\alias{length}
\alias{reads}
\alias{reads1}
\alias{reads2}
\alias{setReads1}
\alias{setReads2}
\alias{show}
\title{reads1}
\usage{
reads1(object, ...)

reads2(object, ...)

setReads1(object, r1)

setReads2(object, r2)
}
\arguments{
\item{reads}{object}

\item{reads}{object}

\item{reads}{object}

\item{r1}{GRangesList object, the new reads to set on the reads object}

\item{reads}{object}

\item{r2}{GRangesList object, the new reads to set on the reads object}
}
\value{
GRangesList The reads of the "+" strand

GRangesList The reads of the "-" strand

reads object

reads object
}
\description{
reads1

reads2

setReads1

setReads2
}

