% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reads1}
\alias{reads1}
\title{reads1}
\usage{
reads1(object)
}
\arguments{
\item{object}{reads object}
}
\value{
GRangesList The reads of the "+" strand
}
\description{
reads1
}
\details{
This method return a GRangesList with the reads corresponding to the "+" strand
}
\seealso{
\code{\link{reads2}}
}

