% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-reads.R, R/methods-segvis.R
\docType{methods}
\name{readsR}
\alias{readsR}
\alias{readsR,reads-method}
\alias{readsR,segvis-method}
\alias{readsR<-}
\alias{readsR<-,reads,list-method}
\alias{readsR<-,segvis,list-method}
\title{readsR methods}
\usage{
readsR(object)

readsR(object) <- value

\S4method{readsR}{reads}(object)

\S4method{readsR}{reads,list}(object) <- value

\S4method{readsR}{segvis}(object)

\S4method{readsR}{segvis,list}(object) <- value
}
\arguments{
\item{object}{Either a \code{segvis} o \code{reads} object}

\item{value}{A list of data.table with same format as a GenomicRanges object with an additional match column}
}
\value{
A list of data.table with same format as a GenomicRanges object with an additional match column

A reads or a segvis object with modified backward reads
}
\description{
readsR returns a list of backward reads of segvis or reads objects

readsR<- assisgn a list of backward reads to a reads or to a segvis object
}
\examples{
\dontrun{
readsR(segvis)
readsR(reads)
readsR(segvis) <- new_reads
}
}
\seealso{
\code{\link{reads-class}},\code{\link{segvis-class}}
}

