% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R, R/methods-segvis_block.R
\docType{methods}
\name{regions}
\alias{regions}
\alias{regions,segvis-method}
\alias{regions,segvis_block-method}
\alias{regions<-}
\alias{regions<-,segvis,GRanges-method}
\alias{regions<-,segvis_block,GRanges-method}
\title{regions methods}
\usage{
regions(object)

regions(object) <- value

\S4method{regions}{segvis}(object)

\S4method{regions}{segvis,GRanges}(object) <- value

\S4method{regions}{segvis_block}(object)

\S4method{regions}{segvis_block,GRanges}(object) <- value
}
\arguments{
\item{object}{A \code{segvis} object}

\item{value}{A GenomicRanges object}
}
\value{
A GenomicRanges object

A segvis object
}
\description{
regions return a GenomicRanges object with the regions used by segvis

regions<- assigns a GenomicRanges object as the regions used by segvis
}
\examples{
\dontrun{

regions(segvis)
regions(segvis) <- gr  ## gr is a GenomicRanges object

}
}
\seealso{
\code{\link{segvis-class}}
}

