% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{regions}
\alias{regions}
\title{regions}
\usage{
regions(object)
}
\arguments{
\item{object}{Either a profile or profileMatrix object}
}
\value{
GRangesList or GRanges depending on the call for profile or profileMatrix object
}
\description{
Generic method for both profile and profileMatrix classes
}
\details{
This methods returns the regions for which the profile and profileMatrix are calculated as a
GRangesList for profile or GRanges for profileMatrix
}
\seealso{
\code{\link{setRegions}},\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

