% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remChr}
\alias{remChr}
\title{remChr method for profile class}
\usage{
remChr(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
character value. Indicates the chromosomes to be ignored
}
\description{
Returns the chromosomes to be ignored
}
\seealso{
\code{\link{profile-class}}, \code{\link{Profile}} and \code{\link{setRemChr}}
}

