% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{segvis-class}
\alias{segvis-class}
\title{Segvis class description}
\description{
This object is the base class of the segvis package. It contains all the information necessary for the calculation of coverage curves.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character with the name of the profiles}

\item{\code{regions}}{GRanges object with the regions for which the profile want to be calcualted}

\item{\code{file}}{Character with the name of the file that contains the reads}

\item{\code{maxBandwidth}}{Numeric value with the maximum bandwidth accepted when smoothing profiles. Must be odd}

\item{\code{fragLen}}{Numeric value with the fragment length to resize the reads (if it is zero then it doesn't resize the reads)}

\item{\code{chr}}{character vector, with the chromosomes to be considered}

\item{\code{isPET}}{logical, Indicates is the reads come from a PET experiments or a SET experiment}

\item{\code{reads}}{Reads object, which contains the reads used to build the profile separated by strand}

\item{\code{profiles}}{RleList - For each region, there is a Rle object}

\item{\code{.haveRegions}}{logical - Indicates if the object have the regions loaded}

\item{\code{.haveReads}}{logical - Indicates if the object have the reads loaded}

\item{\code{.readsMatched}}{logical - Indicates if the read have been matched to the regions}

\item{\code{.coverageCalculated}}{logical - Indicates if the coverage has been calculated]}
}}
\seealso{
\code{\link{buildSegvis}}
}

