% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{countReads}
\alias{countReads}
\title{countReads method for segvis class}
\usage{
countReads(object)
}
\arguments{
\item{object}{segvis object}
}
\value{
The number of reads in the bam file considered for the \code{segvis} object
}
\description{
Counts the number of reads considered  in object
}
\seealso{
\code{\link{segvis}}
}

