% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{joinProfiles}
\alias{joinProfiles}
\title{joinProfiles method for segvis class}
\usage{
joinProfiles(object, bw, mc)
}
\arguments{
\item{object}{segvis object}

\item{bw}{Numeric values with the bandwidth used to smooth the profiles, must be and odd number less or equal than the maxBandwidth}

\item{mc}{Numeric value representing the number of cores used with parallel}
}
\value{
data.table object
}
\description{
Joins all the coverages into a data.table with columns chr|match|coord|tagCounts
}
\examples{
\dontrun{ joinProfiles(segvis_object,bw = 151,mc=8)}
}

