% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{loadReads}
\alias{loadReads}
\title{loadReads method for segvis class}
\usage{
loadReads(object, mc)
}
\arguments{
\item{object}{segvis object}

\item{mc}{numeric, the number of cores used with parallel}
}
\value{
profile object
}
\description{
Load the fragment stored in the file slot of the segvis object. The reads are divided by chromosome and by strand
}
\examples{
\dontrun{ loadReads(profile_object,mc=8)}
}
\seealso{
\code{\link{readsF}}, \code{\link{readsR}} and \code{\link{reads-class}}
}

