% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{matchReads}
\alias{matchReads}
\title{matchReads methods for segvis class}
\usage{
matchReads(object, mc)
}
\arguments{
\item{object}{segvis object}

\item{mc}{numeric, the number of cores used with parallel}
}
\value{
segvis object
}
\description{
Match the reads to the extended regions stored in the object slots.
}
\examples{
\dontrun{ matchReads(profile_object,mc=8)}
}
\seealso{
\code{\link{matchF}}, \code{\link{matchR}} and \code{\link{match-class}}
}

