% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Segvis-package.R
\docType{package}
\name{Segvis-package}
\alias{Segvis-package}
\title{This package has an objective}
\description{
This package has an objective
}
\examples{
## see the package vignette: --- in creation
}
\author{
Rene Welch
}

