% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{segvis_block-class}
\alias{segvis_block-class}
\title{segvis_block class description}
\description{
Contains a data.table with the individual coverages for each region
}
\details{
\section{Slots}{
  \describe{
   \item{\code{name}:}{Character with the name of the profiles}
   \item{\code{regions}:}{GRanges object with the regions of the individuals profiles}
   \item{\code{cover}:}{table data.table with the individuals coverages for each region}
   \item{\code{bandwidth}:}{Numeric value used to smooth the individual profiles}
   \item{\code{normCont}:}{Numeric normalizing contanst}
   \item{\code{.isScaled}:}{Logical value representing if the profiles are scaled to \code{normConst}}
   }
}
}
\seealso{
\code{\link{Segvis_block}}
}

