% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{Segvis_block}
\alias{Segvis_block}
\title{Segvis_block constructor from a segvis object}
\usage{
Segvis_block(object, bw, mc)
}
\arguments{
\item{object}{segvis object}

\item{bw}{Numeric value with the bandwidth used to smooth the coverage profiles, must be and odd number less or equal than the maxBandwidth}

\item{mc}{Numeric value with the number of cores used with parallel}
}
\value{
a segvis_block object
}
\description{
This method returns a segvis_block object built out of the arguments in a segvis object
}
\examples{
\dontrun{

mc <- 4
segvis_block <- Segvis_block(segvis,101,mc)

}
}
\seealso{
\code{\link{segvis_block-class}},\code{\link{segvis-class}}
}

