% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{segvis_block_list-class}
\alias{segvis_block_list-class}
\title{segvis_block_list class description}
\description{
Contains a list of \code{segvis_block} objects
}
\details{
\section{Slots}{
  \describe{
   \item{\code{list}:}{A list of \code{segvis_block} objects or several \code{segvis_block} objects separated by commas}
   }
}
}

