% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segvis_block_list.R
\name{Segvis_block_list}
\alias{Segvis_block_list}
\title{Create a segvis_block_list object}
\usage{
Segvis_block_list(...)
}
\arguments{
\item{...}{Object of the segvis_block class}
}
\value{
A segvis_block_list object
}
\description{
Constructor for the segvis_block_list class
}
\seealso{
\code{\link{segvis_block-class}}
}

