% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Segvis_blockList.R
\name{Segvis_blockList}
\alias{Segvis_blockList}
\title{Create a segvis_blockList object}
\usage{
Segvis_blockList(...)
}
\arguments{
\item{...}{Object of the segvis_block class}
}
\value{
A segvis_blockList object
}
\description{
Constructor for the segvis_blockList class
}
\seealso{
\code{\link{segvis_block-class}}
}

