% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ProfileMatrixList.R, R/SegvisList.R
\name{SegvisList}
\alias{SegvisList}
\title{Create a segvisList object}
\usage{
SegvisList(...)

SegvisList(...)
}
\arguments{
\item{...}{Object of the segvis class}

\item{...}{Object of the segvis class}
}
\value{
A segvisList object

A segvisList object
}
\description{
Constructor for the segvisList class

Constructor for the segvisList class
}
\seealso{
\code{\link{segvis-class}}

\code{\link{segvis-class}}
}

