% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segviz.R
\docType{package}
\name{Segviz}
\alias{Segviz}
\alias{Segviz-package}
\title{A package for visualization of high throughput sequencing data along
genomic segments}
\description{
Segviz is an R package that automatizes visualization tasks for high
throughput sequencing data. It allows the user to plot several marks
accross specific genomic regions and functional profiles around specific
coordinates. Additionally, it is capable of performing these tasks
considering both strands by separate or the aggregate profiles of both
strands.
}
\details{
Segvis provides different tools to summarize and visualize data:
}
