% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_segvis.R
\name{separate.by.chrom}
\alias{separate.by.chrom}
\title{.separate.by.chrom}
\usage{
separate.by.chrom(greads, chrom, st, mc, sort = FALSE)
}
\arguments{
\item{greads}{A data.table with at least the columns seqnames, start, end and strand}

\item{chrom}{A vector with all the chromosome used to split the table}

\item{st}{A character that represents the strand: "+", "-" or "*";
 for forward, reverse or any of them, respectively}

\item{mc}{A numeric value indicating the number of cores}

\item{sort}{Optional boolean to see if the intervals are going to be sorted}
}
\value{
list with greads separated
}
\description{
Separates a set of reads by chromosome while filters by strand (and
have the posibility to sort)
}

