% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setBandwidth}
\alias{setBandwidth}
\title{setBandwidth method for profileMatrix class}
\usage{
setBandwidth(object, newBandwidth)
}
\arguments{
\item{object}{profileMatrix object}

\item{newBandwidth}{Numeric value with the bandwidth used to smooth the profile in profileMatrix object}
}
\value{
profileMatrix object
}
\description{
Set method for bandwidth parameter in the profileMatrix class
}
\seealso{
\code{\link{bandwidth}}
}

