% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setFragLen}
\alias{setFragLen}
\title{setFragLen method for profile class}
\usage{
setFragLen(object, newFragLen)
}
\arguments{
\item{object}{profile object}

\item{newFragLen}{Numeric value, must be greater of equal to zero}
}
\value{
profile object
}
\description{
Set method for the fragLen parameter
}
\seealso{
\code{\link{profile-class}}, \code{\link{Profile}} and \code{\link{fragLen}}
}

