% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setMaxBandwidth}
\alias{setMaxBandwidth}
\title{setMaxBandwidth method for profile class}
\usage{
setMaxBandwidth(object, newMaxBandwidth)
}
\arguments{
\item{object}{profile object}

\item{newMaxBandwidth}{Numeric value, must be odd and greater or equal than one}
}
\value{
profile object
}
\description{
Set method for the maxBandwidth parameter
}
\seealso{
\code{\link{profile-class}}, \code{\link{Profile}} and \code{\link{maxBandwidth}}
}

