% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setNormConst}
\alias{setNormConst}
\title{setNormConst method for profileMatrix class}
\usage{
setNormConst(object, newNormConst)
}
\arguments{
\item{object}{profileMatrix object}

\item{newNormConst}{Numeric value with the constant used to normalize the profile}
}
\value{
profileMatrix object
}
\description{
Set method for a normalizing constant in the profileMatrix class
}
\seealso{
\code{\link{normConst}}
}

