% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setProfileMat}
\alias{setProfileMat}
\title{setProfileMat method for profileMatrix class}
\usage{
setProfileMat(object, newProfileMat)
}
\arguments{
\item{object}{profileMatrix object}

\item{newProfileMat}{A matrix with a the profile for each region}
}
\value{
profileMatrix object
}
\description{
Set method for the profileMat parameter of the profileMatrix class
}
\seealso{
\code{\link{profileMat}}
}

