% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setReads2}
\alias{setReads2}
\title{setReads2 method for reads class}
\usage{
setReads2(object, r2)
}
\arguments{
\item{object}{reads object}

\item{r2}{GRangesList object, the new reads to set on the reads object}
}
\value{
reads object
}
\description{
Set method for the reads class, set the reads in the reverse strand
}
\seealso{
\code{\link{reads1}},\code{\link{reads2}} and \code{\link{setReads1}}
}

