% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{method}
\name{setRemChr}
\alias{setRemChr}
\title{setRemChr method for profile class}
\usage{
setRemChr(object, newRemChr)
}
\arguments{
\item{object}{profile object}

\item{newRemChr}{Character vector, with the chromosomes to be removed}
}
\value{
profile object
}
\description{
Set method for the remChr parameter
}
\seealso{
\code{\link{profile-class}}, \code{\link{Profile}} and \code{\link{remChr}}
}

