% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sort_by_strand}
\alias{sort_by_strand}
\title{Sort the reads on a GRanges object depending on it's strand}
\usage{
sort_by_strand(gr, stra)
}
\arguments{
\item{gr}{- A GRanges object}

\item{stra}{- Unique strand of the object}
}
\value{
A GRanges object
}
\description{
Sort the reads on a GRanges object depending on it's strand
}

