% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis_block.R
\docType{methods}
\name{summarize}
\alias{summarize}
\alias{summarize,ANY-method}
\title{summarize method for segvis_block class}
\usage{
summarize(object, FUN, ...)
}
\arguments{
\item{object}{segvis_block object}

\item{FUN}{Function used to summarize the tagCounts, for example if \code{FUN = mean}, then it will returns a vector with the mean of all tagCounts by centered coordinate. In particular, \code{FUN} needs to take vector argument + optional parameters}

\item{...}{Rest of the arguments needed to use \code{FUN}}
}
\value{
A numeric vector
}
\description{
This methods summarizes the coverage of the data when all the regions have the same width. To do so, it pilles up all tagCounts and applies FUN to each coordinate.
}
\examples{
\dontrun{
summarize(segvis_block,mean)
summarize(segvis_block,mean,trim=.25)
summarize(segvis_block,median)
}
}
\seealso{
\code{\link{segvis_block-class}}
}

