% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{bandwidth}
\alias{bandwidth}
\title{bandwidth method for profileMatrix class}
\usage{
bandwidth(object)
}
\arguments{
\item{object}{profileMatrix class}
}
\value{
numeric. The bandwidth used to smooth the profile in profileMatrix object
}
\description{
Get method for the bandwidth parameter for profileMatrix class
}
\seealso{
\code{\link{setBandwidth}}
}

