% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{buildProfileMatrix}
\alias{buildProfileMatrix}
\title{buildProfileMatrix method for profile class}
\usage{
buildProfileMatrix(object, bw, mc)
}
\arguments{
\item{object}{profile object}

\item{bw,}{the bandwidth used to smooth the profiles, must be and odd number less or equal than the maxBandwidth}

\item{mc,}{the number of cores used with parallel}
}
\value{
matrix object
}
\description{
If all the regions have the same width, this method build a profile matrix of dimension nr. regions x width
}
\examples{
\dontrun{ buildProfileMatrix(profile_object,bw = 151,mc=8)}
}

