% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{bwd_covers}
\alias{bwd_covers}
\alias{bwd_covers,SegvizData-method}
\title{bwd_covers method}
\usage{
bwd_covers(object)

\S4method{bwd_covers}{SegvizData}(object)
}
\arguments{
\item{object}{A \code{SegvizData} object.}
}
\value{
A list of the coverages for all the reads aligned to the reverse
strand.
}
\description{
bwd_covers returns a list with the \code{coverage} of each file in the
\code{SegvizData} object, calculated exclusively with the backward strand
reads.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
bwd_covers(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

