% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{bwd_covers}
\alias{bwd_covers}
\alias{bwd_covers,SegvizData-method}
\alias{bwd_covers}
\title{bwd_covers method}
\usage{
bwd_covers(object)

\S4method{bwd_covers}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{bwd_covers} method returns a list of aggregated coverage
for all the aligned reads files considered, by using the reads in the reverse
strand exclusively.
}
\description{
bwd_covers returns a list with the \code{coverage} of each file in the
\code{SegvizData} object, calculated exclusively with the backward strand
reads.
}
\seealso{
\code{\link{SegvizData-class}}
}
