% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{DT_region}
\alias{DT_region}
\alias{DT_region,SegvizData,GRanges-method}
\alias{DT_region,SegvizData,numeric-method}
\title{DT_region method}
\usage{
DT_region(object, region, ...)

\S4method{DT_region}{SegvizData,GRanges}(object, region,
  nameFiles = basename(files(object)), type = "aggr", normalize = TRUE,
  base = 1e+06)

\S4method{DT_region}{SegvizData,numeric}(object, region,
  nameFiles = basename(files(object)), type = "aggr", normalize = TRUE,
  base = 1e+06)
}
\arguments{
\item{object}{a \code{SegvizData} object.}

\item{region}{a numeric value indicating which region to plot or a
\code{GRanges} object indicating a whole region to plot.}

\item{...}{Any other additional parameters that \code{DT_region} may need.}

\item{nameFiles}{a character vector with the shortened names that are going
to be used in the plot.}

\item{type}{a character value indicating which strand to use. By default
uses the aggregate coverage between both strands.}

\item{normalize}{a logical value indicating if the profiles are normalized
respect to its number of reads.}

\item{base}{a numeric value indicating the number of aligned reads to which
the signal is going to be normalized. The default value is \code{1e6}.}
}
\value{
The \code{DT_region} method returns a \code{data.table} object with
the info. necessary to be plotted.
}
\description{
DT_region returns a \code{data.table} with the info. necessary to plot
the coverage of the aligned read files across the
\code{peak_id}th element of the \code{SegvizData} object.
}
\seealso{
\code{\link{SegvizData-class}}

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
DT_region(segviz, region = 1)
DT_region(segviz,
  region = GRanges(seqnames = "chr2",ranges = IRanges(start = 56,width = 1e3)))
}

