% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fileFormat}
\alias{fileFormat}
\title{fileFormat method for profile class}
\usage{
fileFormat(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
character. A character with the file format of the file where reads are stored
}
\description{
Returns the format of the file where the reads are stored
}
\seealso{
\code{\link{profile-class}}
}

