% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fileFormat}
\alias{fileFormat}
\title{fileFormat}
\usage{
fileFormat(object, ...)
}
\arguments{
\item{object}{A profile object}
}
\value{
character. A character with the file format of the file where reads are stored
}
\description{
Returns the format of the file where the reads are stored
}
\details{
This method return the value of the fileFormat slot
}
\seealso{
\code{\link{profile-class}}
}

