% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvisData.R
\docType{methods}
\name{files}
\alias{files}
\alias{files,SegvisData-method}
\alias{files}
\title{files method}
\usage{
files(object)

\S4method{files}{SegvisData}(object)
}
\arguments{
\item{object}{a \code{SegvisData} object.}
}
\value{
The \code{files} method returns a character vector with the files
used in the \code{SegvisData} object.
}
\description{
files returns a character vectors with the files used in the SegvisData
object.
}
\seealso{
\code{\link{SegvisData-class}}
}
