% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{files}
\alias{files}
\alias{files,SegvizData-method}
\title{files method}
\usage{
files(object)

\S4method{files}{SegvizData}(object)
}
\arguments{
\item{object}{A \code{SegvizData} object.}
}
\value{
A character vector with files used in the \code{SegvizData} object.
}
\description{
files returns a character vectors with the files used in the SegvizData
object.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
files(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

