% Generated by roxygen2 (4.0.1): do not edit by hand
\name{.subset_profileMat_logical}
\alias{.filter_profileMat}
\alias{.subset_profileMat_logical}
\title{Set of functions to subset profileMatrix objects}
\usage{
.subset_profileMat_logical(object, condition_call)

.filter_profileMat(object, cond)
}
\arguments{
\item{object}{profileMatrix object}

\item{condition_call}{name object with the filtering condition}

\item{cond}{Logical Vector with the peaks that satisfy certain condition}
}
\value{
A logical vector with the same length as the number of peaks

A profileMat object with the peaks that such that cond is TRUE
}
\description{
\code{.subset_profileMat_logical} Function to obtain a logical vector of the peaks in regions(object) such that satisfy condition_all

\code{.filter_profileMat} Filters the profileMatrix object using a logical vector as an input, the cond vector can be obtained by using \code{.subset_profileMat_logical}
}

