% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{find_summits}
\alias{find_summits}
\alias{find_summits,SegvizData-method}
\title{find_summits method}
\usage{
find_summits(object, ...)

\S4method{find_summits}{SegvizData}(object, which.file = 1,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{object}{A \code{SegvizData} object.}

\item{which.file}{A numeric value indicating which file is going to be
used to calculate the summits. By default it takes the first file.}

\item{mc.cores}{A numeric value with the number of cores to use,
i.e. at most how many child processes will be run simultaneously.}
}
\value{
A numeric vector with the genomic coordinate with the summit in the
\code{GRanges} according to one of the files in object.
}
\description{
find_summits returns a numeric vector with the same length of the
\code{GRanges} with the summits according to one of the files.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
segviz$summit = find_summits(segviz,which.file = 1 ,mc.cores = 2)

}
\seealso{
\code{\link{SegvizData-class}}
}

