% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{findSummit}
\alias{findSummit}
\title{findSummit method for profile class}
\usage{
findSummit(object, bw, mc)
}
\arguments{
\item{object}{profile object}

\item{bw,}{the bandwidth used to smooth the profiles, must be and odd number less or equal than the maxBandwidth}

\item{mc,}{the number of cores used with parallel}
}
\value{
Numeric vector
}
\description{
This method finds the summit of each region whenever is possible, otherwise returns NA
}
\examples{
\dontrun{ findSummit(profile_object,bw = 151,mc=8)}
}

