% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{frag_len}
\alias{frag_len}
\alias{frag_len,SegvizData-method}
\title{frag_len method}
\usage{
frag_len(object)

\S4method{frag_len}{SegvizData}(object)
}
\arguments{
\item{object}{A \code{SegvizData} object.}
}
\value{
A numeric vector with the fragment length used to extend the 5' ends
of the the aligned reads. If the reads were paired ended, then a value
of \code{1} is assigned.
}
\description{
frag_len returns a numeric vector with the fragment length used to extend
the 5' ends of the aligned reads.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
frag_len(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

