% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fragLen}
\alias{fragLen}
\title{fragLen method for profile class}
\usage{
fragLen(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
numeric value. This indicated the fragment length
}
\description{
Returns the fragment length value used to extend the reads when matching them with the regions
}
\seealso{
\code{\link{profile-class}}, \code{\link{Profile}} and \code{\link{setFragLen}}
}

