% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{fragLen}
\alias{fragLen}
\title{fragLen}
\usage{
fragLen(object, ...)
}
\arguments{
\item{object}{A profile object}
}
\value{
numeric value. This indicated the fragment length
}
\description{
Returns the fragment length value
}
\details{
This value indicates the length used to extend the fragment when matching reads with regions
}
\seealso{
\code{\link{profile-class}} and \code{\link{Profile}}
}

