% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{fwd_covers}
\alias{fwd_covers}
\alias{fwd_covers,SegvizData-method}
\title{fwd_covers methods}
\usage{
fwd_covers(object)

\S4method{fwd_covers}{SegvizData}(object)
}
\arguments{
\item{object}{A \code{SegvizData} object.}
}
\value{
A list of the coverages for all the reads aligned to the forward
strand.
}
\description{
fwd_covers returns a list with the \code{coverage} of each file in the
\code{SegvizData} object, calculated exclusively with the forward strand
reads.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
fwd_covers(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

