% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{is_pet}
\alias{is_pet}
\alias{is_pet,SegvizData-method}
\title{is_pet method}
\usage{
is_pet(object)

\S4method{is_pet}{SegvizData}(object)
}
\arguments{
\item{object}{A \code{SegvizData} object.}
}
\value{
A logical vector indicating if the aligned reads files are paired
ended.
}
\description{
is_pet returns a logical vector indicating if the aligned reads are paired
end
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
is_pet(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

