% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{is_pet}
\alias{is_pet}
\alias{is_pet,SegvizData-method}
\alias{is_pet}
\title{is_pet method}
\usage{
is_pet(object)

\S4method{is_pet}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{is_pet} method returns a logical vector indicating which
aligned reads files are paired ended.
}
\description{
is_pet returns a logical vector indicating if the aligned reads are paired
end
}
\seealso{
\code{\link{SegvizData-class}}
}
