% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{match-class}
\alias{match-class}
\title{match class description}
\description{
Contains the match of a set a reads of a ChIP - Seq experiment given a set of regions in the genome.
}
\section{Slots}{

\describe{
\item{\code{matchF}}{- List of positions of the reads in list that match with each region and have + strand}

\item{\code{matchR}}{- List of positions of the reads in list that match with each region and have - strand}
}}
\seealso{
\code{\link{matchF}} and  \code{\link{matchR}}
}

