% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{matchF}
\alias{matchF}
\title{matchF method for profile and match classes}
\usage{
matchF(object)
}
\arguments{
\item{match}{object}
}
\value{
In case of the get method returns the reads of the forward strand
}
\description{
Get method for match class with "+" strand
}
\seealso{
\code{\link{match-class}}
}

