% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{maxBandwidth}
\alias{maxBandwidth}
\title{maxBandwidth}
\usage{
maxBandwidth(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
numeric value. This indicated the maximum bandwidth available to later smooth the profiles
}
\description{
Returns the maximum bandwidth value
}
\details{
Return the maxumum bandwidth value
}
\seealso{
\code{\link{profile-class}} and \code{\link{Profile}}
}

