% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{name}
\alias{chr}
\alias{chr<-}
\alias{file}
\alias{file<-}
\alias{fragLen}
\alias{fragLen<-}
\alias{isPET}
\alias{isPET<-}
\alias{maxBandwidth}
\alias{maxBandwidth<-}
\alias{name}
\alias{name<-}
\alias{profiles}
\alias{regions}
\alias{regions<-}
\title{Get and set methods for segvis class}
\usage{
name(object)

name(object) <- value

regions(object)

regions(object) <- value

file(object)

file(object) <- value

maxBandwidth(object)

maxBandwidth(object) <- value

fragLen(object)

fragLen(object) <- value

chr(object)

chr(object) <- value

isPET(object)

isPET(object) <- value

profiles(object)

\S4method{name}{segvis}(object)

\S4method{regions}{segvis}(object)

\S4method{file}{segvis}(object)

\S4method{maxBandwidth}{segvis}(object)

\S4method{fragLen}{segvis}(object)

\S4method{chr}{segvis}(object)

\S4method{isPET}{segvis}(object)

\S4method{profiles}{segvis}(object)

\S4method{name}{segvis,character}(object) <- value

\S4method{regions}{segvis,GRanges}(object) <- value

\S4method{file}{segvis,character}(object) <- value

\S4method{maxBandwidth}{segvis,numeric}(object) <- value

\S4method{fragLen}{segvis,numeric}(object) <- value

\S4method{chr}{segvis,character}(object) <- value

\S4method{isPET}{segvis,logical}(object) <- value
}
\arguments{
\item{object}{An object of class segvis}

\item{value}{The class of this parameter depends on the function, for \code{file<-}, \code{name<-} and \code{chr<-} is a string; for \code{fragLen<-} and  \code{maxBandwidth} is a numeric value; for \code{regions<-} is a GRanges object and for \code{isPET<-} is a logical value}

\item{object}{A segvis object}
}
\value{
Returns the value with the same name as the function

Returns a segvis object with the value indicated by the functions being updated with value

RleList. A list made of an Rle object for each region
}
\description{
This methods determine how to get and set attributes dor the segvis class. The structure is similar to the one use by bioconductor objects

Returns the coverage for each region considered
}
\seealso{
\code{\link{segvis-class}}

\code{\link{segvis-class}} and \code{\link{segvis-getCoverage}}
}

