% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{name}
\alias{name}
\title{name method for profile and profileMatrix}
\usage{
name(object)
}
\arguments{
\item{object}{Either a profile or profileMatrix object}
}
\value{
character The name of the object
}
\description{
Generic method for both profile and profileMatrix classes that returns the name of the object
}
\seealso{
\code{\link{setName}},\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

