% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{normConst}
\alias{normConst}
\title{normConst method for profileMatrix class}
\usage{
normConst(object)
}
\arguments{
\item{object}{profileMatrix object}
}
\value{
numeric. The constant used to normalize the profile
}
\description{
Get method for the normConst parameter for profileMatrix class
}
\seealso{
\code{\link{setNormConst}}
}

