% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{nreads}
\alias{nreads}
\alias{nreads,SegvizData-method}
\title{nreads methods}
\usage{
nreads(object)

\S4method{nreads}{SegvizData}(object)
}
\arguments{
\item{object}{A \code{SegvizData} object.}
}
\value{
A numeric vector with the depth of each experiment.
}
\description{
nreads returns a numeric vector with the number of the depth of all the
experiments considered.
}
\examples{

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
nreads(segviz)

}
\seealso{
\code{\link{SegvizData-class}}
}

