% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{nreads}
\alias{nreads}
\alias{nreads,SegvizData-method}
\alias{nreads}
\title{nreads method}
\usage{
nreads(object)

\S4method{nreads}{SegvizData}(object)
}
\arguments{
\item{object}{a \code{SegvizData} object.}
}
\value{
The \code{nreads} method returns a numeric vector with the depth
of each experiment.
}
\description{
nreads returns a numeric vector with the number of the depth of all the
experiments considered.
}
\seealso{
\code{\link{SegvizData-class}}
}
