% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvisData.R
\docType{methods}
\name{plot_region}
\alias{plot_region}
\alias{plot_region,SegvisData,GRanges-method}
\alias{plot_region}
\alias{plot_region,SegvisData,numeric-method}
\alias{plot_region}
\title{plot_region method}
\usage{
plot_region(object, region, ...)

\S4method{plot_region}{SegvisData,GRanges}(object, region,
  nameFiles = basename(files(object)), type = "aggr", normalize = TRUE,
  base = 1e+06)

\S4method{plot_region}{SegvisData,numeric}(object, region,
  nameFiles = basename(files(object)), type = "aggr", normalize = TRUE,
  base = 1e+06)
}
\arguments{
\item{object}{a \code{SegvisData} object.}

\item{region}{a numeric value indicating which region to plot in the
\code{SegvisData} object or a \code{GRanges} object with a genomic region
that is not recorded in the \code{SegvisData} object}

\item{...}{Any other additional parameters that \code{plot_region} may need.}

\item{nameFiles}{a character vector with the shortened names that are going
to be used in the plot.}

\item{type}{a character value indicating which strand to use. By default
uses the aggregate coverage between both strands.}

\item{normalize}{a logical value indicating if the profiles are normalized
respect to its number of reads.}

\item{base}{a numeric value indicating the number of aligned reads to which
the signal is going to be normalized. The default value is \code{1e6}.}
}
\value{
The \code{plot_region} method returns a \code{ggplot} object with
the desired plot.
}
\description{
plot_region plots the coverage of the aligned read files across the
\code{peak_id}th element of the \code{SegvisData} object.
}
\seealso{
\code{\link{SegvisData-class}}
}
