% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-SegvizData.R
\docType{methods}
\name{plot_region}
\alias{plot_region}
\alias{plot_region,SegvizData,numeric-method}
\title{plot_region method}
\usage{
plot_region(object, peak_id, ...)

\S4method{plot_region}{SegvizData,numeric}(object, peak_id,
  nameFiles = basename(files(object)), type = "aggr", normalize = TRUE,
  base = 1e+06)
}
\arguments{
\item{object}{a \code{SegvizData} object.}

\item{peak_id}{a numeric value indicating which region to plot.}

\item{nameFiles}{a character vector with the shortened names that are going
to be used in the plot.}

\item{type}{a character value indicating which strand to use. By default
uses the aggregate coverage between both strands.}

\item{normalize}{a logical value indicating if the profiles are normalized
respect to its number of reads.}

\item{base}{a numeric value indicating the number of aligned reads to which
the signal is going to be normalized. The default value is \code{1e6}.}
}
\value{
The \code{plot_region} method returns a \code{ggplot} object with
the desired plot.
}
\description{
plot_region plots the coverage of the aligned read files across the
\code{peak_id}th element of the \code{SegvizData} object.
}
\seealso{
\code{\link{SegvizData-class}}

## load SegvizData object
load(system.file("extdata","example","segviz.RData",package = "Segvis"))
plot_region(segviz, peak_id = 1)
}

