% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remChr}
\alias{match}
\alias{reads}
\alias{remChr}
\title{remChr}
\usage{
remChr(object, ...)

reads(object)

match(object, ...)
}
\arguments{
\item{profile}{object}

\item{profile}{object}

\item{profile}{object}
}
\value{
character. A character vector representing the chromosomes to be ignored

reads. A reads object with the reads contained in file

match object, with the index in reads of the matching reads for the regions
}
\description{
remChr

reads

match
}

