% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Profile}
\alias{Profile}
\title{Creates a profile objec}
\usage{
Profile(regionName, file, fileFormat, maxBandwidth, fragLen, remChr = NULL)
}
\arguments{
\item{regionName}{Name of the profile object}

\item{file}{Character string with the file name that contain the reads}

\item{fileFormat}{Character vector with the file format used}

\item{maxBandwidth}{Numeric value of the maximum possible value to smooth the profile curve}

\item{fragLen}{Numeric value of the fragment length used to extend the reads}

\item{remChr}{Character Vector with the chromosomes to be ignored}
}
\value{
A profile object
}
\description{
Constructor for profile class
}
\examples{
rn = "sites"
f = "somefile.bam";fileF = "bam"
maxBw = 1;fl = 200
Profile(regionName = rn,file =f,fileFormat=fileF,maxBandwidth = maxBw, fragLen = fl)
}

