% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProfileMatrix}
\alias{ProfileMatrix}
\title{Create a profileMatrix object}
\usage{
ProfileMatrix(prof, bw, mc = 8)
}
\arguments{
\item{prof}{- Profile object}

\item{bw}{- Numeric value representing the bandwidth used to smooth the profiles}

\item{mc}{- Numeric value representing the number of cores used to run in parallel mode}
}
\value{
a profileMatrix object
}
\description{
Create a profileMatrix object
}

