% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProfileMatrix}
\alias{ProfileMatrix}
\title{Create a profileMatrix object}
\usage{
ProfileMatrix(name, regions)
}
\arguments{
\item{name}{Character with the name of the profileMatrix object}

\item{regions}{GRanges object with the regions considered for the profile, the width of all regions must be the same}
}
\value{
a profileMatrix object
}
\description{
Constructor for profileMatrix class
}

