% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProfileMatrixList}
\alias{ProfileMatrixList}
\title{Create a profileMatrixList object}
\usage{
ProfileMatrixList(...)
}
\arguments{
\item{...}{Objects of the profileMatrix class}
}
\value{
A profileMatrixList object
}
\description{
Constructor for the profileMatrixList class
}
\seealso{
\code{\link{profileMatrixList-class}}
}

