% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{reads-class}
\alias{reads-class}
\title{reads class description}
\description{
Contains the reads obtained in a ChIP - seq experiment
}
\section{Slots}{

\describe{
\item{\code{readsF}}{List of data.table objects containing the reads of the
ChIP - Seq experiment that have + strand.}

\item{\code{readsR}}{List of data.table object containning the reads of the
ChIP - Seq experiment that have - strand}
}}
\seealso{
\code{\link{readsF}}, \code{\link{readsR}} and \code{\link{loadReads}}
}

