% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{reads-class}
\alias{reads-class}
\title{reads class}
\description{
Contains the reads obtained in a ChIP - seq experiment separated by strand and then by chromosome.
}
\section{Slots}{

\describe{
\item{\code{reads1}}{- GRangesList of the reads of the ChIP - Seq experiment that have + strand}

\item{\code{reads2}}{- GRangesList of the reads of the ChIP - Seq experiment that have - strand}
}}

