% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reads1}
\alias{reads1}
\title{Generic method reads1 for profile and reads classes}
\usage{
reads1(object)
}
\arguments{
\item{object}{reads object}
}
\value{
GRangesList The reads of the "+" strand
}
\description{
Get methods for reads class with "+" strand
}
\seealso{
\code{\link{reads2}}
}

