% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{reads2}
\alias{reads2}
\title{Generic method reads2 for profile and reads classes}
\usage{
reads2(object)
}
\arguments{
\item{object}{reads object}
}
\value{
GRangesList The reads of the "-" strand
}
\description{
Get methods for reads class with "-" strand
}
\seealso{
\code{\link{reads1}}
}

