% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-reads.R, R/methods-segvis.R
\docType{methods}
\name{readsF}
\alias{readsF}
\alias{readsF<-}
\title{readsF methods}
\usage{
readsF(object)

readsF(object) <- value

\S4method{readsF}{reads}(object)

\S4method{readsF}{reads,list}(object) <- value

\S4method{readsF}{segvis}(object)
}
\arguments{
\item{object}{Either a \code{segvis} or a \code{reads} object}

\item{value}{A list of data.table with same format as a GenomicRanges object with an additional match column}
}
\value{
A list of data.table with same format as a GenomicRanges object with an additional match column

A reads or a segvis object with modified forward reads
}
\description{
readsF returns a list of forward reads of segvis or reads objects

readsF<- assisgn a list of forward reads to a reads or to a segvis object
}
\examples{
\dontrun{
readsF(segvis)
readsF(reads)
readsF(segvis) <- new_reads
}
}
\seealso{
\code{\link{reads-class}},\code{\link{segvis-class}}
}

