% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{readsF}
\alias{readsF}
\title{Generic method readsF for profile and reads classes}
\usage{
readsF(object)
}
\arguments{
\item{object}{reads object}
}
\value{
Returns a GRangesList with the reads of the forward strand
}
\description{
Get method for reads and profile classes that returns the reads from the forward strand
}
\seealso{
\code{\link{reads-class}}
}

