% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{readsR}
\alias{readsR}
\title{Generic method readsR for profile and reads classes}
\usage{
readsR(object)
}
\arguments{
\item{object}{reads object}
}
\value{
Returns a GRangesList with the reads of the backward strand
}
\description{
Get method for reads and profile classes that returns the reads from the backward strand
}
\seealso{
\code{\link{reads-class}}
}

