% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{regions}
\alias{regions}
\alias{regions<-}
\title{regions methods for profile and profileMatrix}
\usage{
regions(object)

regions(object) <- value
}
\arguments{
\item{object}{Either a profile or profileMatrix object}

\item{value}{GRanges object with the regions for which the profiles are calculated}
}
\value{
GRanges object with the regions for which the profiles are calculated

profile object with the regions slot replaced by the GRanges object value
}
\description{
Generic set and get methods for both profile and profileMatrix classes that returns the regions for which the profiles are calculated
}
\seealso{
\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

