% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{regions}
\alias{regions}
\title{regions methods for profile and profileMatrix}
\usage{
regions(object)
}
\arguments{
\item{object}{Either a profile or profileMatrix object}
}
\value{
GRanges object with the regions for which the profiles are calculated
}
\description{
Generic method for both profile and profileMatrix classes that returns the regions for which the profiles are calculated
}
\seealso{
\code{\link{setRegions}},\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

