% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{remChr}
\alias{remChr}
\title{remChr}
\usage{
remChr(object)
}
\arguments{
\item{object}{A profile object}
}
\value{
character value. Indicates the chromosomes to be ignored
}
\description{
Returns the chromosomes to be ignored
}
\details{
This value indicater the chromosomes that should be ignored
}
\seealso{
\code{\link{profile-class}} and \code{\link{Profile}}
}

