% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Segvis.R
\name{Segvis}
\alias{Segvis}
\title{Creates a segvis objec}
\usage{
Segvis(regionName, file, maxBandwidth, chr, fragLen = 0, isPET = FALSE)
}
\arguments{
\item{regionName}{Name of the segvis object}

\item{file}{Character string with the file name that contain the reads}

\item{maxBandwidth}{Numeric value of the maximum possible value to smooth the profile curve}

\item{chr}{String or string vector specifying the organism or chromosomes to be used}

\item{fragLen}{Numeric value of the fragment length used to extend the reads}
}
\value{
A segvis object
}
\description{
Constructor for segvis class
}
\section{Slots}{

\describe{
\item{\code{isPET}}{logical, Indicates is the reads come from a PET experiments or a SET experiment}
}}

