% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segvis.R
\docType{package}
\name{Segvis}
\alias{Segvis}
\alias{Segvis-package}
\title{A package for visualization of high throughput sequencing data along
genomic segments}
\description{
Segvis is an R package that automatizes visualization tasks for high
throughput sequencing data. It allows the user to plot several marks
accross specific genomic regions and functional profiles around specific
coordinates.
}
\details{
Segvis provides different tools to summarize and visualize data:
}

