% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segvis}
\alias{segvis}
\title{\code{SegvisData} example of histone marks in CTCF peaks}
\format{SegvisData object}
\usage{
data(segvis)
}
\value{
A \code{SegvisData} object
}
\description{
\code{SegvisData} segvis, generated by calling \code{SegvisData}
}
