% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segvis.R
\docType{package}
\name{Segvis}
\alias{Segvis}
\alias{Segvis-package}
\title{A package for visualization of high throughout sequencing data along genomic segments}
\description{
Segvis is an R package which can automatize tasks related to the
visualization of aligned read data.
}
\details{
Segvis provides different tools to summarize and visualize data:

\itemize{
 \item Extract read data of specified input regions

\item Plot peaks of different conditions across the same set of regions

\item Plot coordinate-wise statistics for a set of same width regions

\item Subset regions according to used defined annotations and conditions

}
}

