% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-segvis.R
\docType{methods}
\name{Segvis_block}
\alias{Segvis_block}
\title{Create a segvis_block object}
\usage{
Segvis_block(object, bw, mc)
}
\arguments{
\item{object}{segvis object}

\item{bw}{Numeric value with the bandwidth used to smooth the coverage profiles, must be and odd number less or equal than the maxBandwidth}

\item{mc}{Numeric value with the number of cores used with parallel}
}
\value{
a sevis_block object
}
\description{
Constructor for segvis_block class, it takes both name and region from \code{object}
}
\seealso{
\code{\link{segvis_block-class}}
}

