% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SegvizData-class}
\alias{SegvizData}
\alias{SegvizData-class}
\title{SegvizData object and constructors}
\usage{
SegvizData(regions, files, is_pet = rep(FALSE, length(files)), frag_len = 1,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{regions}{a \code{GRanges} object with the regions to be considered.}

\item{files}{a character vector with the location of the bam files that
contain the aligned reads.}

\item{is_pet}{a logical vector with the same length as files indicating if
the aligned reads in the respective bam file are paired.}

\item{frag_len}{a numeric vector representing the average fragment length
to which the aligned reads in their respective bam file are going to be
extended. For PE reads, this parameter is not considered.}

\item{mc.cores}{A numeric value with the number of cores to use,
i.e. at most how many child processes will be run simultaneously.}
}
\value{
\code{SegvizData} returns a \code{SegvizData} object which contains
the forward, backward and aggregated coverage for every aligned read file in
\code{files}.
}
\description{
\code{SegvizData} is a subclass of \code{GenomicRanges}, used to visualize
high-thoughput sequencing experiments across a set of user - defined
genomic regions.
}
\examples{
dr = system.file("inst","extdata","example",package = "Segvis")
files = list.files(dr,pattern = "bam",full.names =TRUE)
files = files[grep("bai",files,invert = TRUE)][1:3]

reg = list.files(dr,pattern = "narrow",full.names =TRUE)
reg = readBedFile(reg[1])
segviz = SegvizData(regions = reg,files)

}

