% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setName}
\alias{setName}
\title{setName method for profile and profileMatrix classes}
\usage{
setName(object, newName)
}
\arguments{
\item{object}{Either a profile or profileMatrix object}

\item{newName}{character}
}
\value{
profile or profileMatrix object
}
\description{
Generic set name method for both profile and profileMatrix classes
}
\seealso{
\code{\link{name}},\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

