% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setReads1}
\alias{setReads1}
\title{setReads1 method for reads class}
\usage{
setReads1(object, r1)
}
\arguments{
\item{object}{reads object}

\item{r1}{GRangesList object, this are the new reads to be set in the reads object}
}
\value{
reads object
}
\description{
Set method of the reads class, set the reads in the forward strand
}
\seealso{
\code{\link{reads1}},\code{\link{reads2}} and \code{\link{setReads2}}
}

