% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setReads2}
\alias{setReads2}
\title{setReads2}
\usage{
setReads2(object, r2)
}
\arguments{
\item{reads}{object}

\item{r2}{GRangesList object, the new reads to set on the reads object}
}
\value{
reads object
}
\description{
Set method for the reads class, for the reads with "-" strand
}
\details{
This method return a reads object where the reads with "-" strand have been modified by the user
}
\seealso{
\code{\link{reads1}},\code{\link{reads2}} and \code{\link{setReads1}}
}

