% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setRegions}
\alias{setRegions}
\title{setRegions}
\usage{
setRegions(object, newRegions)
}
\arguments{
\item{object}{Either a profile or profileMatrix object}

\item{newRegions}{GRangesList for profile or GRanges for profileMatrix}
}
\value{
profile object
}
\description{
Generic set regions method for both profile and profileMatrix classes
}
\details{
This methods returns either a profile or profileMatrix object
}
\seealso{
\code{\link{regions}},\code{\link{profile-class}} and \code{\link{profileMatrix-class}}
}

